initial commit for WF extension
This is just an initial PoC to get things wired up. There is no configuration
yet. The HTTP server is bound to port 9090 on localhost. This will eventually
need to be configurable through socket bindings.

standalone.xml has to be manually configured as well for now. The extension
needs to be declared,

<extensions>
  ...
  <extension module="org.hawkular.metrics.server"/>
</extension>

and the subsystem needs to be declared,

<subsystem xmlns="urn:org.hawkular.metrics:server:1.0"/>