fix race condition in DataAccessImpl.insertGaugeTag
There was a race condition that could result in the batch statement to be empty
which means no data would get inserted into the tags table. Here is a pseudo
code description of the problem,

//---------------------------------------------------------------------------
BatchStatement batch = // new batch statement
data.forEach(...)  // data is the Gauge observable and the function passed to
                   // forEach was creating and adding statements to the batch
rxSession.execute(batch)
//---------------------------------------------------------------------------

The problem was that the batch statement was getting updated asynchronously.
The data observable is async, which means the call to execute() could happen
before the batch was fully updated. I have refactored the method making it
fully async. The batch statement is populated now via a reduce operation and
then executed in a flatMap operation.