initial integration with H-Alerts for stand alone deployment
The dist module produces a WildFly 10 distro that includes H-Metrics and
H-Alerts. It comes with the messaging subsystem pre-configured.

There is an initial integration test in the dist-itests module that submits data
points via the REST API, listens on the HawkularMetricData topic, and verifies
that it receives a notification via JMS.

The hawkular-component module has been refactored a bit to use JMS 2.0 instead
of the hawkular-bus-common APIs. I also updated MetricDataPublisher to send
Gauge objects in the JMS message. There are two reasons for this. First, Gauge
and other classes in hawkular-metrics-api-common define the model of the public
API, so it makes sense to use those classes when and where possible. Secondly,
we can avoid some extra JSON marshalling that we are currently doing. This can
be achievd by writing the Gauge objects to the topic when they get unmarshalled
rahter than first unmarshalling them to their core model counterparts first.